Rename areas view to areas overview view
