+++ src/panels/lovelace/strategies/areas/helpers/areas-strategy-helper.ts
@@ -4,6 +4,8 @@ import type { EntityFilterFunc } from "../../../../../common/entity/entity_filte
 import { generateEntityFilter } from "../../../../../common/entity/entity_filter";
 import { stripPrefixFromEntityName } from "../../../../../common/entity/strip_prefix_from_entity_name";
 import { orderCompare } from "../../../../../common/string/compare";
+import type { AreaRegistryEntry } from "../../../../../data/area_registry";
+import { areaCompare } from "../../../../../data/area_registry";
 import type { LovelaceCardConfig } from "../../../../../data/lovelace/config/card";
 import type { HomeAssistant } from "../../../../../types";
 import { supportsAlarmModesCardFeature } from "../../../card-features/hui-alarm-modes-card-feature";
@@ -251,3 +253,25 @@ export const computeAreaTileCardConfig =
       ...additionalCardConfig,
     };
   };
+
+export const getAreas = (
+  entries: HomeAssistant["areas"],
+  hiddenAreas?: string[],
+  areasOrder?: string[]
+): AreaRegistryEntry[] => {
+  const areas = Object.values(entries);
+
+  const filteredAreas = hiddenAreas
+    ? areas.filter((area) => !hiddenAreas!.includes(area.area_id))
+    : areas.concat();
+
+  const compare = areaCompare(entries, areasOrder);
+
+  const sortedAreas = filteredAreas.sort((areaA, areaB) =>
+    compare(areaA.area_id, areaB.area_id)
+  );
+
+  return sortedAreas;
+};
+
+export const computeAreaPath = (areaId: string): string => `areas-${areaId}`;
