+++ src/panels/lovelace/strategies/get-strategy.ts
@@ -33,7 +33,7 @@ const STRATEGIES: Record<LovelaceStrategyConfigType, Record<string, any>> = {
     map: () => import("./map/map-view-strategy"),
     iframe: () => import("./iframe/iframe-view-strategy"),
     area: () => import("./areas/area-view-strategy"),
-    areas: () => import("./areas/areas-view-strategy"),
+    "areas-overview": () => import("./areas/areas-overview-view-strategy"),
   },
   section: {},
 };
